# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/visitor_tracks:
    get:
      tags:
        - VisitorTracks
      summary: List VisitorTracks
      description: Get a paginated list of VisitorTracks with optional filtering and sorting
      operationId: listVisitorTracks
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitorTrackList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - VisitorTracks
      summary: Create VisitorTrack
      description: Create a new VisitorTrack
      operationId: createVisitorTrack
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVisitorTrackRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitorTrack'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/website_audit_logs:
    get:
      tags:
        - WebsiteAuditLogs
      summary: List WebsiteAuditLogs
      description: Get a paginated list of WebsiteAuditLogs with optional filtering and sorting
      operationId: listWebsiteAuditLogs
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebsiteAuditLogList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - WebsiteAuditLogs
      summary: Create WebsiteAuditLog
      description: Create a new WebsiteAuditLog
      operationId: createWebsiteAuditLog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWebsiteAuditLogRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebsiteAuditLog'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    VisitorTrack:
      type: object
      required:
        - id
        - visitor_id
        - url
        - visited_at
      properties:
        id:
          type: string
          format: uuid
        visitor_id:
          type: string
          format: uuid
        page_id:
          nullable: true
          type: string
          format: uuid
        url:
          type: string
          maxLength: 500
        visited_at:
          type: string
          format: date-time
...
    CreateVisitorTrackRequest:
      type: object
      required:
        - visitor_id
        - url
        - visited_at
      properties:
        visitor_id:
          type: string
          format: uuid
        page_id:
          nullable: true
          type: string
          format: uuid
        url:
          type: string
          maxLength: 500
        visited_at:
          type: string
          format: date-time
...
    VisitorTrackList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/VisitorTrack'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    WebsiteAuditLog:
      type: object
      required:
        - id
        - event
        - occurred_at
      properties:
        id:
          type: string
          format: uuid
        event:
          type: string
        actor:
          nullable: true
          type: string
          format: uuid
        subject_type:
          nullable: true
          type: string
          maxLength: 120
        subject_id:
          nullable: true
          type: string
          format: uuid
        detail:
          nullable: true
          type: object
        occurred_at:
          type: string
          format: date-time
...
    CreateWebsiteAuditLogRequest:
      type: object
      required:
        - event
        - occurred_at
      properties:
        event:
          type: string
        actor:
          nullable: true
          type: string
          format: uuid
        subject_type:
          nullable: true
          type: string
          maxLength: 120
        subject_id:
          nullable: true
          type: string
          format: uuid
        detail:
          nullable: true
          type: object
        occurred_at:
          type: string
          format: date-time
...
    WebsiteAuditLogList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/WebsiteAuditLog'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []